pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "aiolimiter>=1.2.1",
    "fastapi>=0.129.0",
    "gunicorn>=25.1.0",
    "jinja2>=3.1.6",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "scipy>=1.17.0",
    "structlog>=25.5.0",
    "uvicorn>=0.41.0",
    "websockets>=16.0",
]
